Refactor Chembl extraction service layering
