+++ src/bioetl.egg-info/SOURCES.txt
@@ -32,7 +32,7 @@ src/bioetl/application/pipelines/chembl/factories.py
 src/bioetl/application/pipelines/chembl/pipeline.py
 src/bioetl/application/pipelines/chembl/transformer.py
 src/bioetl/application/services/__init__.py
-src/bioetl/application/services/chembl_extraction.py
+src/bioetl/infrastructure/clients/chembl/impl/chembl_extraction_service_impl.py
 src/bioetl/domain/__init__.py
 src/bioetl/domain/contracts.py
 src/bioetl/domain/enums.py
