+++ src/bioetl/infrastructure/clients/chembl/impl/__init__.py
@@ -2,3 +2,9 @@
 ChEMBL implementation details.
 """
 
+from bioetl.infrastructure.clients.chembl.impl.chembl_extraction_service_impl import (
+    ChemblExtractionServiceImpl,
+)
+
+__all__ = ["ChemblExtractionServiceImpl"]
+
